@@ -12,11 +12,14 @@
     <url>http://topjava.herokuapp.com/</url>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>15</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-        <spring.version>5.3.3</spring.version>
+        <spring.version>5.3.4</spring.version>
+        <spring-data-jpa.version>2.4.5</spring-data-jpa.version>
+        <jackson-json.version>2.12.2</jackson-json.version>
+        <tomcat.version>9.0.43</tomcat.version>
 
         <!-- Logging -->
         <logback.version>1.2.3</logback.version>
