@@ -1,10 +1,21 @@
 package ru.javawebinar.topjava.model;
 
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
+
+import javax.persistence.Column;
+import javax.persistence.MappedSuperclass;
+
+
+@MappedSuperclass
 public abstract class AbstractNamedEntity extends AbstractBaseEntity {
 
+    @NotBlank
+    @Size(min = 2, max = 100)
+    @Column(name = "name", nullable = false)
     protected String name;
 
-    public AbstractNamedEntity() {
+    protected AbstractNamedEntity() {
     }
 
     protected AbstractNamedEntity(Integer id, String name) {
