@@ -7,6 +7,7 @@ import ru.javawebinar.topjava.repository.UserRepository;
 
 import java.util.Comparator;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 import static ru.javawebinar.topjava.UserTestData.admin;
@@ -18,8 +19,9 @@ public class InMemoryUserRepository extends InMemoryBaseRepository<User> impleme
 
     public void init() {
         map.clear();
-        map.put(UserTestData.USER_ID, user);
-        map.put(UserTestData.ADMIN_ID, admin);
+        put(user);
+        put(admin);
+        counter.getAndSet(UserTestData.ADMIN_ID + 1);
     }
 
     @Override
@@ -31,6 +33,7 @@ public class InMemoryUserRepository extends InMemoryBaseRepository<User> impleme
 
     @Override
     public User getByEmail(String email) {
+        Objects.requireNonNull(email, "email must not be null");
         return getCollection().stream()
                 .filter(u -> email.equals(u.getEmail()))
                 .findFirst()
